requirements.txt
python-docx==0.8.11
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.0
click==8.1.7
PyYAML==6.0.1
requests==2.31.0
colorama==0.4.6
tqdm==4.66.1
watchdog==3.0.0 